requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9

passlib[bcrypt]==1.7.4
pyotp==2.9.0
python-jose==3.3.0
qrcode==7.4.2
Pillow==10.4.0
